refactor(UtilService): Move removeHTMLTags #1085
